Removed old shim files
